+++ components/pages/home/homeList.tsx
@@ -1,9 +1,9 @@
 import Link from "next/link";
 
-import styles from "../../../styles/pages/home/homeList.module.scss";
+import homeListStyles from "@styles/pages/home/homeList.module.scss";
 
 import { dateFormatter } from "@/utils/dateFormatter";
-import { ContentDataType } from "@/components/pages/contents/contentList";
+import { ContentDataType } from "@/components/pages/contents/contents";
 import { RefObject } from "react";
 import CustomImage from "@/components/common/customImage";
 
@@ -15,7 +15,7 @@ export default function HomeList({
   listRef: RefObject<HTMLUListElement>;
 }) {
   return (
-    <ul ref={listRef} className={styles.home_list}>
+    <ul ref={listRef} className={homeListStyles.home_list}>
       {listData.map((val: ContentDataType) => {
         const { id, poster_path, vote_average } = val;
 
@@ -27,21 +27,23 @@ export default function HomeList({
           <li key={id}>
             <Link href={`/contentDetail?type=${type}&id=${id}`}>
               <CustomImage
-                className={styles.image}
+                className={homeListStyles.image}
                 type="content"
                 src={`https://image.tmdb.org/t/p/w220_and_h330_face/${poster_path}`}
               />
             </Link>
-            <div className={styles.vote_average}>{vote}</div>
-            <div className={styles.title_release}>
+            <div className={homeListStyles.vote_average}>{vote}</div>
+            <div className={homeListStyles.title_release}>
               <Link
-                className={styles.title}
+                className={homeListStyles.title}
                 href={`/contentDetail?type=${type}&id=${id}`}
               >
                 <div>{val.name || val.title}</div>
               </Link>
 
-              <div className={styles.release}>{dateFormatter(date)}</div>
+              <div className={homeListStyles.release}>
+                {dateFormatter(date)}
+              </div>
             </div>
           </li>
         );
