+++ components/pages/home/searchBar.tsx
@@ -1,7 +1,7 @@
 import { KeyboardEvent, useState } from "react";
 import { useRouter } from "next/navigation";
 
-import styles from "../../../styles/pages/home/searchBar.module.scss";
+import searchBarStyles from "@styles/pages/home/searchBar.module.scss";
 
 export default function SearchBar() {
   const router = useRouter();
@@ -18,15 +18,15 @@ export default function SearchBar() {
   };
 
   return (
-    <div className={styles.search_bar}>
+    <div className={searchBarStyles.search_bar}>
       <div>
-        <div className={styles.intro_text}>
+        <div className={searchBarStyles.intro_text}>
           <div>Welcome.</div>
           <div>
             Millions of movies, TV shows and people to discover. Explore now.
           </div>
         </div>
-        <div className={styles.search_input}>
+        <div className={searchBarStyles.search_input}>
           <input
             value={inputValue}
             placeholder="영화, TV 프로그램, 인물 검색..."
