+++ components/pages/home/trendingList.tsx
@@ -3,9 +3,9 @@ import { useEffect, useRef, useState } from "react";
 import { apiClient } from "@/api/httpClient";
 import HomeFilterBar from "./categoryTab";
 import HomeList from "./homeList";
-import { ContentDataType } from "@/components/pages/contents/contentList";
+import { ContentDataType } from "@/components/pages/contents/contents";
 
-import styles from "../../../styles/pages/home/contents.module.scss";
+import contentsStyles from "@styles/pages/home/contents.module.scss";
 
 export default function TrendingList({ list }: any) {
   const listRef = useRef<HTMLUListElement>(null);
@@ -39,8 +39,8 @@ export default function TrendingList({ list }: any) {
   }, []);
 
   return (
-    <div className={styles.content}>
-      <div className={styles.title_bar}>
+    <div className={contentsStyles.content}>
+      <div className={contentsStyles.title_bar}>
         <h2>트렌딩</h2>
         <HomeFilterBar
           tabData={tabData}
